Flytter generering av utbetalingsoppdrag inn i egen service for å unngå sirkulær avhengighet
