+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/økonomi/ØkonomiService.kt
@@ -1,24 +1,13 @@
 package no.nav.familie.ba.sak.integrasjoner.økonomi
 
 import io.micrometer.core.instrument.Metrics
-import no.nav.familie.ba.sak.common.toYearMonth
-import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiUtils.grupperAndeler
-import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiUtils.oppdaterBeståendeAndelerMedOffset
-import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
-import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
-import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
-import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
 import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseValideringService
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseRepository
 import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelseRepository
 import no.nav.familie.ba.sak.kjerne.beregning.domene.utbetalingsoppdrag
 import no.nav.familie.ba.sak.kjerne.simulering.KontrollerNyUtbetalingsgeneratorService
 import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
-import no.nav.familie.felles.utbetalingsgenerator.domain.BeregnetUtbetalingsoppdragLongId
-import no.nav.familie.felles.utbetalingsgenerator.domain.IdentOgType
 import no.nav.familie.http.client.RessursException
 import no.nav.familie.kontrakter.felles.oppdrag.OppdragId
 import no.nav.familie.kontrakter.felles.oppdrag.OppdragStatus
@@ -26,20 +15,15 @@ import no.nav.familie.kontrakter.felles.oppdrag.Utbetalingsoppdrag
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-import java.time.YearMonth
 
 @Service
 class ØkonomiService(
-    private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService,
     private val økonomiKlient: ØkonomiKlient,
     private val beregningService: BeregningService,
-    private val utbetalingsoppdragGenerator: UtbetalingsoppdragGenerator,
-    private val behandlingService: BehandlingService,
     private val tilkjentYtelseValideringService: TilkjentYtelseValideringService,
     private val tilkjentYtelseRepository: TilkjentYtelseRepository,
-    private val andelTilkjentYtelseRepository: AndelTilkjentYtelseRepository,
     private val kontrollerNyUtbetalingsgeneratorService: KontrollerNyUtbetalingsgeneratorService,
+    private val utbetalingsgeneratorService: UtbetalingsgeneratorService,
 
 ) {
     private val sammeOppdragSendtKonflikt = Metrics.counter("familie.ba.sak.samme.oppdrag.sendt.konflikt")
@@ -50,7 +34,7 @@ class ØkonomiService(
         andelTilkjentYtelseForUtbetalingsoppdragFactory: AndelTilkjentYtelseForUtbetalingsoppdragFactory,
     ): Utbetalingsoppdrag {
         val oppdatertBehandling = vedtak.behandling
-        val utbetalingsoppdrag = genererUtbetalingsoppdragOgOppdaterTilkjentYtelse(
+        val utbetalingsoppdrag = utbetalingsgeneratorService.genererUtbetalingsoppdragOgOppdaterTilkjentYtelse(
             vedtak,
             saksbehandlerId,
             andelTilkjentYtelseForUtbetalingsoppdragFactory,
@@ -97,147 +81,6 @@ class ØkonomiService(
         } else {
             OppdragStatus.KVITTERT_OK
         }
-
-    fun genererUtbetalingsoppdrag(
-        vedtak: Vedtak,
-        saksbehandlerId: String,
-        erSimulering: Boolean = false,
-    ): BeregnetUtbetalingsoppdragLongId {
-        val forrigeBehandling =
-            behandlingHentOgPersisterService.hentForrigeBehandlingSomErIverksatt(behandling = vedtak.behandling)
-        val forrigeTilkjentYtelse =
-            forrigeBehandling?.let { tilkjentYtelseRepository.findByBehandlingAndHasUtbetalingsoppdrag(behandlingId = it.id) }
-        val nyTilkjentYtelse = tilkjentYtelseRepository.findByBehandling(behandlingId = vedtak.behandling.id)
-        val endretMigreringsDato = beregnOmMigreringsDatoErEndret(
-            vedtak.behandling,
-            forrigeTilkjentYtelse?.andelerTilkjentYtelse?.minByOrNull { it.stønadFom }?.stønadFom,
-        )
-        val sisteAndelPerKjede =
-            andelTilkjentYtelseRepository.hentSisteAndelPerIdent(fagsakId = vedtak.behandling.fagsak.id)
-                .associateBy { IdentOgType(it.aktør.aktivFødselsnummer(), it.type.tilYtelseType()) }
-        return utbetalingsoppdragGenerator.lagUtbetalingsoppdrag(
-            saksbehandlerId = saksbehandlerId,
-            vedtak = vedtak,
-            forrigeTilkjentYtelse = forrigeTilkjentYtelse,
-            nyTilkjentYtelse = nyTilkjentYtelse,
-            sisteAndelPerKjede = sisteAndelPerKjede,
-            erSimulering = erSimulering,
-            endretMigreringsDato = endretMigreringsDato,
-        )
-    }
-
-    @Transactional
-    fun genererUtbetalingsoppdragOgOppdaterTilkjentYtelse(
-        vedtak: Vedtak,
-        saksbehandlerId: String,
-        andelTilkjentYtelseForUtbetalingsoppdragFactory: AndelTilkjentYtelseForUtbetalingsoppdragFactory,
-        erSimulering: Boolean = false,
-        skalValideres: Boolean = true,
-    ): Utbetalingsoppdrag {
-        val oppdatertBehandling = vedtak.behandling
-        val oppdatertTilstand =
-            beregningService.hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(behandlingId = oppdatertBehandling.id)
-                .pakkInnForUtbetaling(andelTilkjentYtelseForUtbetalingsoppdragFactory)
-
-        val oppdaterteKjeder = grupperAndeler(oppdatertTilstand)
-
-        val erFørsteIverksatteBehandlingPåFagsak =
-            beregningService.hentTilkjentYtelseForBehandlingerIverksattMotØkonomi(fagsakId = oppdatertBehandling.fagsak.id)
-                .isEmpty()
-
-        val utbetalingsoppdrag = if (erFørsteIverksatteBehandlingPåFagsak) {
-            utbetalingsoppdragGenerator.lagUtbetalingsoppdragOgOppdaterTilkjentYtelse(
-                saksbehandlerId = saksbehandlerId,
-                vedtak = vedtak,
-                erFørsteBehandlingPåFagsak = erFørsteIverksatteBehandlingPåFagsak,
-                oppdaterteKjeder = oppdaterteKjeder,
-                erSimulering = erSimulering,
-            )
-        } else {
-            val forrigeBehandling =
-                behandlingHentOgPersisterService.hentForrigeBehandlingSomErIverksatt(behandling = oppdatertBehandling)
-                    ?: error("Finner ikke forrige behandling ved oppdatering av tilkjent ytelse og iverksetting av vedtak")
-
-            val forrigeTilstand =
-                beregningService.hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(forrigeBehandling.id)
-                    .pakkInnForUtbetaling(andelTilkjentYtelseForUtbetalingsoppdragFactory)
-
-            val forrigeKjeder = grupperAndeler(forrigeTilstand)
-
-            val sisteOffsetPerIdent = beregningService.hentSisteOffsetPerIdent(
-                forrigeBehandling.fagsak.id,
-                andelTilkjentYtelseForUtbetalingsoppdragFactory,
-            )
-
-            val sisteOffsetPåFagsak = beregningService.hentSisteOffsetPåFagsak(behandling = oppdatertBehandling)
-
-            if (oppdatertTilstand.isNotEmpty()) {
-                oppdaterBeståendeAndelerMedOffset(oppdaterteKjeder = oppdaterteKjeder, forrigeKjeder = forrigeKjeder)
-                val tilkjentYtelseMedOppdaterteAndeler = oppdatertTilstand.first().tilkjentYtelse
-                beregningService.lagreTilkjentYtelseMedOppdaterteAndeler(tilkjentYtelseMedOppdaterteAndeler)
-            }
-
-            val utbetalingsoppdrag = utbetalingsoppdragGenerator.lagUtbetalingsoppdragOgOppdaterTilkjentYtelse(
-                saksbehandlerId = saksbehandlerId,
-                vedtak = vedtak,
-                erFørsteBehandlingPåFagsak = erFørsteIverksatteBehandlingPåFagsak,
-                forrigeKjeder = forrigeKjeder,
-                sisteOffsetPerIdent = sisteOffsetPerIdent,
-                sisteOffsetPåFagsak = sisteOffsetPåFagsak,
-                oppdaterteKjeder = oppdaterteKjeder,
-                erSimulering = erSimulering,
-                endretMigreringsDato = beregnOmMigreringsDatoErEndret(
-                    vedtak.behandling,
-                    forrigeTilstand.minByOrNull { it.stønadFom }?.stønadFom,
-                ),
-            )
-
-            if (!erSimulering && (
-                    oppdatertBehandling.type == BehandlingType.MIGRERING_FRA_INFOTRYGD_OPPHØRT || behandlingHentOgPersisterService.hent(
-                        oppdatertBehandling.id,
-                    ).resultat == Behandlingsresultat.OPPHØRT
-                    )
-            ) {
-                utbetalingsoppdrag.validerOpphørsoppdrag()
-            }
-
-            utbetalingsoppdrag
-        }
-
-        if (skalValideres) {
-            utbetalingsoppdrag.validerNullutbetaling(
-                behandlingskategori = vedtak.behandling.kategori,
-                andelerTilkjentYtelse = beregningService.hentAndelerTilkjentYtelseForBehandling(vedtak.behandling.id),
-            )
-        }
-
-        opprettAdvarselLoggVedForstattInnvilgetMedUtbetaling(utbetalingsoppdrag, vedtak.behandling)
-
-        return utbetalingsoppdrag
-    }
-
-    private fun beregnOmMigreringsDatoErEndret(behandling: Behandling, forrigeTilstandFraDato: YearMonth?): YearMonth? {
-        val erMigrertSak =
-            behandlingHentOgPersisterService.hentBehandlinger(behandling.fagsak.id)
-                .any { it.type == BehandlingType.MIGRERING_FRA_INFOTRYGD }
-
-        if (!erMigrertSak) {
-            return null
-        }
-
-        val nyttTilstandFraDato = behandlingService.hentMigreringsdatoPåFagsak(fagsakId = behandling.fagsak.id)
-            ?.toYearMonth()
-            ?.plusMonths(1)
-
-        return if (forrigeTilstandFraDato != null &&
-            nyttTilstandFraDato != null &&
-            forrigeTilstandFraDato.isAfter(nyttTilstandFraDato)
-        ) {
-            nyttTilstandFraDato
-        } else {
-            null
-        }
-    }
 
     companion object {
