+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/simulering/SimuleringServiceEnhetTest.kt
@@ -8,6 +8,7 @@ import no.nav.familie.ba.sak.common.randomFnr
 import no.nav.familie.ba.sak.common.tilPersonEnkel
 import no.nav.familie.ba.sak.config.FeatureToggleConfig
 import no.nav.familie.ba.sak.config.FeatureToggleService
+import no.nav.familie.ba.sak.integrasjoner.økonomi.UtbetalingsgeneratorService
 import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiKlient
 import no.nav.familie.ba.sak.integrasjoner.økonomi.ØkonomiService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
@@ -47,15 +48,16 @@ internal class SimuleringServiceEnhetTest {
     private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService = mockk()
     private val persongrunnlagService: PersongrunnlagService = mockk()
     private val kontrollerNyUtbetalingsgeneratorService: KontrollerNyUtbetalingsgeneratorService = mockk()
+    private val utbetalingsgeneratorService: UtbetalingsgeneratorService = mockk()
 
     private val simuleringService: SimuleringService = SimuleringService(
         økonomiKlient,
-        økonomiService,
         beregningService,
         økonomiSimuleringMottakerRepository,
         tilgangService,
         featureToggleService,
         vedtakRepository,
+        utbetalingsgeneratorService,
         behandlingHentOgPersisterService,
         persongrunnlagService,
         kontrollerNyUtbetalingsgeneratorService,
